Completed Spring 14

+++ src/main/java/ru/practicum/shareit/item/ItemDao.java
@@ -2,7 +2,6 @@ package ru.practicum.shareit.item;
 
 
 import org.springframework.stereotype.Component;
-import ru.practicum.shareit.item.model.Item;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -11,9 +10,9 @@ import java.util.Map;
 
 @Component
 public class ItemDao {
-    private Map<Integer, Item> itemMap = new HashMap<>();
+    private Map<Long, Item> itemMap = new HashMap<>();
 
-    public Item getById(int id) {
+    public Item getById(Long id) {
         return itemMap.get(id);
     }
